03_ssai_chatbot/multi_agent_chatbot/requirements.txt
# Multi-Agent Chatbot Dependencies

# Core AI/ML libraries
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20

# LangGraph for workflow management
langgraph>=0.0.20
langchain-core>=0.1.0

# MCP (Model Context Protocol) support
mcp>=1.0.0
langchain-mcp-adapters>=0.1.10

# Environment and configuration
python-dotenv>=1.0.0

# Data processing
requests>=2.31.0

# GitHub integration
GitPython>=3.1.40
PyGithub>=1.59.0

# Vector database and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
unstructured>=0.11.0
markdown>=3.4.0

# Web framework (for API server)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Streamlit web interface
streamlit>=1.28.0
streamlit-chat>=0.1.1
streamlit-option-menu>=0.3.6

# Database
sqlalchemy>=2.0.0

# Utilities
pydantic-settings>=2.1.0
typing-extensions>=4.8.0
pathlib2>=2.3.7

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging and monitoring
loguru>=0.7.0

# Additional for multi-agent system
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0
